requirements.txt
bottle==0.13.4
certifi==2026.2.25
charset-normalizer==3.4.4
idna==3.11
proxy_tools==0.1.0
pycairo==1.29.0
PyGObject==3.50.0
PyGObject-stubs==2.16.0
pywebview==6.1
requests==2.32.5
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
